server/pyproject.toml
[tool.poetry]
name = "metro_cms_reset"
version = "0.1.0"
description = ""
authors = ["Виноградов Константин <kostya.06.05.1995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
django-debug-toolbar = "^3.2.2"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
